Refactored API module to be separate modules
