+++ docs/usage/functionality.rst
@@ -1,14 +1,30 @@
 Functionality
 =====================================================
 
-Module: api
+Module: project
 -----------------------------------------------------
 
-.. automodule:: maproulette.api.api
+.. automodule:: maproulette.api.project
     :members:
+    :show-inheritance:
+
+Module: challenge
+-----------------------------------------------------
+
+.. automodule:: maproulette.api.challenge
+    :members:
+    :show-inheritance:
+
+Module: task
+-----------------------------------------------------
+
+.. automodule:: maproulette.api.task
+    :members:
+    :show-inheritance:
 
 Module: configuration
 -----------------------------------------------------
 
 .. automodule:: maproulette.api.configuration
     :members:
+    :show-inheritance:
