+++ examples/add_tasks_to_challenge.py
@@ -5,7 +5,7 @@ import json
 config = maproulette.Configuration(api_key="API_KEY")
 
 # Create an API objects with the above config object:
-api = maproulette.Api(config)
+api = maproulette.Challenge(config)
 
 # Specify the ID of the challenge you want to add tasks to:
 challenge_id = 'TEST_ID'
