+++ examples/get_task_by_id.py
@@ -5,7 +5,7 @@ import json
 config = maproulette.Configuration(api_key="API_KEY")
 
 # Create an API objects with the above config object:
-api = maproulette.Api(config)
+api = maproulette.Task(config)
 
 # Specify the ID of the challenge you want up add tasks to:
 task_id = '42914448'
