+++ maproulette/api/maproulette_server.py
@@ -10,7 +10,26 @@ class MapRouletteServer:
     """Class that holds the basic requests that can be made to the MapRoulette API."""
     def __init__(self, configuration):
         self.url = configuration.url
+        self.base_url = configuration.base_url
         self.headers = configuration.headers
+        if self.check_health():
+            self.session = requests.Session()
+
+    def check_health(self):
+        """Checks health of connection to host by pinging the URL set in the configuration"""
+        response = requests.get(
+            self.base_url + '/ping',
+            headers=self.headers,
+            verify=False
+        )
+        if response.ok:
+            return True
+        else:
+            raise ConnectionUnavailableError(
+                message='Specified server unavailable',
+                status=response.status_code,
+                payload=response
+            )
 
     def get(self, endpoint, params=None):
         """Method that completes a GET request to the MapRoulette API
@@ -19,7 +38,7 @@ class MapRouletteServer:
         :param params: the parameters that pertain to the request (optional)
         :returns: a JSON object containing the API response
         """
-        response = requests.get(
+        response = self.session.get(
             self.url + endpoint,
             params=params,
             headers=self.headers,
@@ -62,7 +81,7 @@ class MapRouletteServer:
         :param body: the body of the request (optional)
         :returns: a JSON object containing the API response
         """
-        response = requests.post(
+        response = self.session.post(
             self.url + endpoint,
             json=body,
             headers=self.headers,
@@ -107,7 +126,7 @@ class MapRouletteServer:
         :param body: the body of the request (optional)
         :returns: a JSON object containing the response code and the API response if
         """
-        response = requests.put(
+        response = self.session.put(
             self.url + endpoint,
             json=body,
             headers=self.headers,
@@ -151,7 +170,7 @@ class MapRouletteServer:
         :param endpoint: the server endpoint to use for the DELETE request
         :returns: a JSON object containing the API response
         """
-        response = requests.delete(
+        response = self.session.delete(
             self.url + endpoint,
             headers=self.headers)
         try:
@@ -186,3 +205,17 @@ class MapRouletteServer:
             return {
                 "status": response.status_code
             }
+
+    @staticmethod
+    def is_json(input_object):
+        """Method to determine whether user input is valid JSON.
+
+        :param input_object: the user's input to check
+        :returns: True if valid json object
+        """
+        try:
+            json_object = json.loads(input_object)
+            del json_object
+            return True
+        except ValueError:
+            return False
