+++ tests/test_task_api.py
@@ -1,0 +1,16 @@
+import maproulette
+import unittest
+from unittest.mock import patch
+
+
+class TestTaskAPI(unittest.TestCase):
+
+    config = maproulette.Configuration(api_key="API_KEY")
+    api = maproulette.Task(config)
+
+    @patch('maproulette.api.maproulette_server.requests.Session.get')
+    def test_get_task_by_id(self, mock_request, api_instance=api):
+        test_task_id = '42914448'
+        mock_request.return_value.status_code = '200'
+        response = api_instance.get_task_by_id(test_task_id)
+        self.assertEqual(response['status'], '200')
